+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed --diskless-on-remaining not working with COMBINED nodes
 - Fixed unnecessary restart of local DRBD peer right after the resource was created
 - ToggleDisk: Added missing free-space check
+- Clone: Added missing free-space check
 
 ## [1.33.0-rc.1] - 2025-11-11